moving shared function to nmlutils
